doc and revise codex verification
