+++ src/codex.rs
@@ -30,15 +30,18 @@ use amplify::Bytes32;
 use commit_verify::{CommitId, CommitmentId, DigestExt, ReservedBytes, Sha256};
 
 use crate::{
-    CellAddr, ContractId, Identity, Instr, IoCat, Operation, StateCell, StateData, StateValue,
-    VerifiedOperation, LIB_NAME_ULTRASONIC,
+    CellAddr, ContractId, Identity, Instr, Operation, StateCell, StateValue, VerifiedOperation,
+    VmContext, LIB_NAME_ULTRASONIC,
 };
 
+/// Identifier of a contract method call.
 pub type CallId = u16;
-pub type AccessId = u16;
 
 /// Codex is a crucial part of a contract; it provides a set of commitments to the contract terms
 /// and conditions expressed as a deterministic program able to run in SONIC computer model.
+///
+/// The main (and the only) operation of the codex is verification of contract [`Operation`]s. It is
+/// done in [`Self::verify`] method.
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
 #[derive(CommitEncode)]
 #[commit_encode(strategy = strict, id = CodexId)]
@@ -46,21 +49,89 @@ pub type AccessId = u16;
 #[strict_type(lib = LIB_NAME_ULTRASONIC)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize), serde(rename_all = "camelCase"))]
 pub struct Codex {
-    pub version: ReservedBytes<2>,
+    /// Consensus version of the codex.
+    ///
+    /// # Future use
+    ///
+    /// For now, the only supported version is one; thus, a `ReservedBytes` is used.
+    ///
+    /// In the future, with more versions coming, this should be replaced with an enum, where the
+    /// first byte will encode (with standard strict encoding) a version number as an enum variant.
+    /// For instance,
+    ///
+    /// ```ignore
+    /// pub enum Codex {
+    ///     V0(CodexV0),
+    ///     V1(CodexV1)
+    /// }
+    /// pub struct CodexV0 { /*...*/ }
+    /// pub struct CodexV1 { /*...*/ }
+    /// ```
+    pub version: ReservedBytes<1>,
+    /// Human-readable name of the codex used in the UI.
     pub name: TinyString,
+    /// Identity of the codex developer.
     pub developer: Identity,
+    /// Timestamp of the codex creation.
+    ///
+    /// This field can be also used to "mine" a vanity codex id. While this feature is noa
+    /// necessary one, many people will try to do it, and it is better to provide them with a
+    /// standard way of doing this, rather to force them into abusing and misusing other fields of
+    /// the codex.
     pub timestamp: i64,
+    /// The order of the field used by VM for all scripts (operation verification and state access
+    /// condition satisfaction).
     pub field_order: u256,
+    /// Input config is used by the VM to verify the satisfaction of the lock conditions for
+    /// operation inputs.
     pub input_config: CoreConfig,
+    /// VM core configuration for the operation verification.
     pub verification_config: CoreConfig,
+    /// List of verifiers for each of the calls supported by the codex.
     pub verifiers: TinyOrdMap<CallId, LibSite>,
-    /// Reserved for the future codex extensions
-    pub reserved: ReservedBytes<8>,
 }
 
 impl Codex {
+    /// The codex id holds a commitment to all codex data.
+    ///
+    /// The codex is encoded using strict encoding into the hasher, which is provided by the
+    /// `#[derive(CommitEncode)]` and `#[commit_encode(strategy = strict, id = CodexId)]` macros in
+    /// the structure definition.
+    ///
+    /// It is the same as the result of the [`CommitId::commit_id`] procedure.
     pub fn codex_id(&self) -> CodexId { self.commit_id() }
 
+    /// The main purpose of the codex is to verify the operation under the contract. This is the
+    /// implementation of this verification procedure.
+    ///
+    /// # Arguments
+    ///
+    /// - `contract_id`: since the contract is external to the codex, this information must be
+    ///   provided to the verifier. While operation also commits to the contract id, this id must
+    ///   come not from the operation itself, but from the external knowledge of the contract id
+    ///   which is being verified; such that operation commitment to the contract is also checked.
+    /// - `operation`: the operation to verify.
+    /// - `memory`: an object holding an actual contract state (see [`Memory`] trait) and provides a
+    ///   read access to it.
+    /// - `repo`: a repository holding VM libraries used in the operation verification, calls to
+    ///   which are kept in the codex (see [`Codex::verifiers`]) _and_ may be called from by the
+    ///   access conditions of the inputs. See [`LibRepo`] for the details.
+    ///
+    /// # Returns
+    ///
+    /// On success, returns a operation wrapped as [`VerifiedOperation`] structure, which should be
+    /// used (1) for updating the contract state by applying the operation, and (2) for the
+    /// persistence of the contract history.
+    ///
+    /// # Errors
+    ///
+    /// On any verification failure, the method does not proceed with further certification and
+    /// instantly returns with one of [`CallError`] variants.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the `repo` (library resolver) returns a library which id doesn't match the
+    /// requested one.
     pub fn verify(
         &self,
         contract_id: ContractId,
@@ -68,7 +139,18 @@ impl Codex {
         memory: &impl Memory,
         repo: &impl LibRepo,
     ) -> Result<VerifiedOperation, CallError> {
-        let resolver = |lib_id: LibId| repo.get_lib(lib_id);
+        let resolver = |lib_id: LibId| {
+            let lib = repo.get_lib(lib_id)?;
+            // We must have this verification to avoid hacking from the client libraries.
+            if lib.lib_id() != lib_id {
+                panic!(
+                    "The library returned by the `LibRepo` provided for the contract operation \
+                     verification doesn't match the requested library id. This error indicates \
+                     that the software using the consensus verification is invalid or compromised."
+                )
+            }
+            Some(lib)
+        };
 
         if operation.contract_id != contract_id {
             return Err(CallError::WrongContract {
@@ -77,12 +159,13 @@ impl Codex {
             });
         }
 
-        // Phase one: get inputs, verify access conditions
+        // Phase 1: get inputs, verify their presence in the memory and access conditions
         let mut vm_inputs = Vm::<aluvm::gfa::Instr<LibId>>::with(self.input_config, GfaConfig {
             field_order: self.field_order,
         });
         let mut read_once_input = SmallVec::new();
         for input in &operation.destroying {
+            // Read memory
             let cell = memory
                 .read_once(input.addr)
                 .ok_or(CallError::NoReadOnceInput(input.addr))?;
@@ -109,15 +192,22 @@ impl Codex {
                 vm_inputs.reset();
             }
 
-            let _ = read_once_input.push(cell.data);
+            // We have same-sized arrays, so we happily skip the result returned by the confined
+            // collection.
+            let _res = read_once_input.push(cell.data);
+            debug_assert!(_res.is_ok());
         }
 
+        // Check that all read values are present in the memory.
         let mut immutable_input = SmallVec::new();
         for addr in &operation.reading {
             let data = memory
                 .immutable(*addr)
                 .ok_or(CallError::NoImmutableInput(*addr))?;
-            let _ = immutable_input.push(data);
+            // We have same-sized arrays, so we happily skip the result returned by the confined
+            // collection.
+            let _res = immutable_input.push(data);
+            debug_assert!(_res.is_ok());
         }
 
         // Phase 2: Verify operation integrity
@@ -147,43 +237,37 @@ impl Codex {
     }
 }
 
+/// The trait, which must be implemented by a client library for a structure providing access to the
+/// valid and most recent contract state, consisting of two parts: *read-once* (also called
+/// *destructible*, or *owned*) and *immutable* (also called *read-only* or *global*).
 pub trait Memory {
+    /// Read a read-once state created by a specific operation read-once output, which is defined as
+    /// a part of [`Operation::destructible`].
     fn read_once(&self, addr: CellAddr) -> Option<StateCell>;
+    /// Read a read-once state created by a specific operation immutable output, which is defined as
+    /// a part of [`Operation::immutable`].
     fn immutable(&self, addr: CellAddr) -> Option<StateValue>;
 }
 
-#[derive(Clone, PartialEq, Eq, Debug)]
-pub struct VmContext<'ctx> {
-    pub read_once_input: &'ctx [StateValue],
-    pub immutable_input: &'ctx [StateValue],
-    pub read_once_output: &'ctx [StateCell],
-    pub immutable_output: &'ctx [StateData],
-}
-
-impl VmContext<'_> {
-    pub fn state_value(&self, cat: IoCat, index: u16) -> Option<StateValue> {
-        match cat {
-            IoCat::IN_RO => self.read_once_input.get(index as usize).copied(),
-            IoCat::IN_AO => self.immutable_input.get(index as usize).copied(),
-            IoCat::OUT_RO => self
-                .read_once_output
-                .get(index as usize)
-                .map(|cell| cell.data),
-            IoCat::OUT_AO => self
-                .immutable_output
-                .get(index as usize)
-                .map(|cell| cell.value),
-        }
-    }
-}
-
+/// The trait providing access to all the VM code libraries used by the contract, in both operation
+/// verification or state access conditions.
 pub trait LibRepo {
+    /// Get a specific library with the provided id.
+    ///
+    /// If the library is not known and this method returns `None`, but the library is called by the
+    /// operation verification or state access script, the verification will fail with
+    /// [`CallError::Script`].
     fn get_lib(&self, lib_id: LibId) -> Option<&Lib>;
 }
 
+/// Contract operation verification errors returned by [`Codex::verify`].
+///
+/// The name of the error type is chose so since the operation "calls" to a contract method, and the
+/// codex verification verifies the integrity of the call.
 #[derive(Copy, Clone, Eq, PartialEq, Debug, Display, Error)]
 #[display(doc_comments)]
 pub enum CallError {
+    /// operation doesn't belong to the currecnt contract.
     #[cfg_attr(
         feature = "baid64",
         display = "operation doesn't belong to the current contract {expected} (operation \
@@ -201,6 +285,7 @@ pub enum CallError {
     /// operation verifier {0} is not present in the codex.
     NotFound(CallId),
 
+    /// operation references read-once memory cell which was not defined.
     #[cfg_attr(
         feature = "baid64",
         display = "operation references read-once memory cell {0} which was not defined."
@@ -211,6 +296,7 @@ pub enum CallError {
     )]
     NoReadOnceInput(CellAddr),
 
+    /// operation references immutable memory cell which was not defined.
     #[cfg_attr(
         feature = "baid64",
         display = "operation references immutable memory cell {0} which was not defined."
@@ -222,17 +308,17 @@ pub enum CallError {
     /// operation references immutable memory cell {0} which was not defined.
     NoImmutableInput(CellAddr),
 
-    /// operation input locking conditions are unsatisfied.
+    /// operation input access conditions are unsatisfied.
     Lock(Option<fe256>),
 
-    /// verification failure {0}
+    /// verification script failure with status code {0}.
     Script(fe256),
 
-    /// verification failure (details are unspecified).
+    /// verification script failure (no status code is returned from the verification script).
     ScriptUnspecified,
 }
 
-/// Unique codex identifier - a commitment to all codex data
+/// Unique codex identifier - a commitment to all the [`Codex`] data.
 #[derive(Wrapper, Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash, Debug, From)]
 #[wrapper(Deref, BorrowSlice, Hex, Index, RangeOps)]
 #[derive(StrictType, StrictDumb, StrictEncode, StrictDecode)]
@@ -253,7 +339,7 @@ impl From<Sha256> for CodexId {
 }
 
 impl CommitmentId for CodexId {
-    const TAG: &'static str = "urn:ubideco:sonic:codex#2024-11-19";
+    const TAG: &'static str = "urn:ubideco:sonic:codex#2025-05-15";
 }
 
 #[cfg(feature = "baid64")]
